Filtering Algo done – search working
